Cambios a la interfaz, repositiorio y antenimientos

+++ GestionDeTaller.Models/Mantenimientos.cs
@@ -12,6 +12,6 @@ namespace GestionDeTaller.Models
 
         public string Descripcion { get; set; }
 
-        public float CostoFijo { get; set; }
+        public double CostoFijo { get; set; }
     }
 }
